chg: [user_management] add user role_management
