+++ var/www/modules/hiddenServices/Flask_hiddenServices.py
@@ -17,8 +17,6 @@ from flask_login import login_required
 from Date import Date
 from HiddenServices import HiddenServices
 
-from Decorator import login_required
-
 # ============ VARIABLES ============
 import Flask_config
 
@@ -236,7 +234,6 @@ def delete_auto_crawler(url):
 # ============= ROUTES ==============
 
 @hiddenServices.route("/crawlers/", methods=['GET'])
-#@login_required(role="ADMIN")
 @login_required
 def dashboard():
     crawler_metadata_onion = get_crawler_splash_status('onion')
